requirements.txt
-i https://pypi.org/simple
annotated-types==0.6.0 ; python_version >= '3.8'
anyio==4.2.0 ; python_version >= '3.8'
async-timeout==4.0.3 ; python_full_version <= '3.11.2'
certifi==2023.11.17 ; python_version >= '3.6'
cffi==1.16.0 ; python_version >= '3.8'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
click==8.1.7 ; python_version >= '3.7'
cryptography==41.0.7
ecdsa==0.18.0 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
exceptiongroup==1.2.0 ; python_version < '3.11'
fastapi==0.108.0
h11==0.14.0 ; python_version >= '3.7'
hiredis==2.3.2
idna==3.6 ; python_version >= '3.5'
pyasn1==0.5.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pydantic==2.5.3
pydantic-core==2.14.6 ; python_version >= '3.7'
pydantic-settings==2.1.0
pyjwt[crypto]==2.8.0
python-dotenv==1.0.0 ; python_version >= '3.8'
python-jose[cryptography]==3.3.0
redis[hiredis]==5.0.1
requests==2.31.0
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
starlette==0.32.0.post1 ; python_version >= '3.8'
typing-extensions==4.9.0 ; python_version >= '3.8'
urllib3==2.1.0 ; python_version >= '3.8'
uvicorn==0.25.0
